requirements.txt
tweepy~=4.12.1
pandas~=1.5.2
s3fs~=2022.11.0
python-dateutil~=2.8.2
requests~=2.28.1
apache-airflow~=2.5.0
CouchDB~=1.2